DynArray : the type of each dimension numbers is provided by the user
For example:
If we want to get a pointer 'ptr' that is :
	int (* ptr)[2][3][5][7];

Originally, we call the function like this:
	int **** ptr = ndimarr(sizeof(int), 4, 2, 3, 5, 7);
And we assume that the type of '2', '3', '5' and '7' is 'int'. If not,
random error may occur. As we use 'va_arg' to obtain the numbers.

Now the type 'int' is provided the users. We can can the fucntion like
this:
	int **** ptr = ndimarr(sizeof(int), 4, "d", 2, 3, 5, 7);